canvodpy/pyproject.toml
[project]
name = "canvodpy"
version = "0.1.0"
description = "Umbrella package for canVOD ecosystem"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Nicolas F. Bader", email = "nicolas.bader@tuwien.ac.at" },
]
keywords = ["GNSS", "VOD", "vegetation", "remote sensing", "geodesy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: GIS",
]

# Core dependencies - sub-packages from workspace
dependencies = [
    "canvod-readers",
    "canvod-aux",
    "canvod-store",
    "canvod-vod",
    "ipykernel>=7.1.0",
    "marimo>=0.9.0",
    "matplotlib>=3.8.0",
    "pint>=0.20",
    "structlog>=23.0",
    "pydantic>=2.0",
    "tqdm>=4.65",
    "tomli>=2.0.0",
    "python-dotenv>=1.0.0",
    "natsort>=8.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.14",
    "ty>=0.0.9",
]

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

# Workspace package sources
# This tells uv where to find the local packages
[tool.uv.sources]
canvod-readers = { workspace = true }
canvod-aux = { workspace = true }
canvod-store = { workspace = true }
canvod-vod = { workspace = true }

[tool.ty]
extend-exclude = [
    "**/docs/*.ipynb",
    "**/docs/notebooks/*.py",
]
